fix bugs

+++ copybitmap.h
@@ -3,7 +3,7 @@
 
 #include "cbmp.h"
 
-unsigned char* extract_rbg_only(BMP* bmp, int size);
+unsigned char* extract_rbg_only(BMP* bmp, unsigned long size);
 void add_bytes_back(BMP* bmp, unsigned char* pixels);
 void destroy(unsigned char* pixels);
